// src/services/configuration-service/models/domain-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Compliance Framework Configuration Service API
 * This is the API for the Compliance Framework Configuration Service.
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DomainAttestation } from "./domain-attestation";
import { DomainControlsAndObjectives } from "./domain-controls-and-objectives";
import { DomainFinding } from "./domain-finding";
import { DomainLink } from "./domain-link";
import { DomainLocalDefinition } from "./domain-local-definition";
import { DomainLogEntry } from "./domain-log-entry";
import { DomainObservation } from "./domain-observation";
import { DomainProperty } from "./domain-property";
import { DomainRisk } from "./domain-risk";
/**
 *
 * @export
 * @interface DomainResult
 */
export interface DomainResult {
  /**
   *
   * @type {Array<DomainLogEntry>}
   * @memberof DomainResult
   */
  assessmentLogEntries?: Array<DomainLogEntry>;
  /**
   *
   * @type {Array<DomainAttestation>}
   * @memberof DomainResult
   */
  attestations?: Array<DomainAttestation>;
  /**
   *
   * @type {string}
   * @memberof DomainResult
   */
  description?: string;
  /**
   *
   * @type {string}
   * @memberof DomainResult
   */
  end?: string;
  /**
   *
   * @type {Array<DomainFinding>}
   * @memberof DomainResult
   */
  findings?: Array<DomainFinding>;
  /**
   *
   * @type {string}
   * @memberof DomainResult
   */
  id?: string;
  /**
   *
   * @type {Array<DomainLink>}
   * @memberof DomainResult
   */
  links?: Array<DomainLink>;
  /**
   *
   * @type {DomainLocalDefinition}
   * @memberof DomainResult
   */
  localDefinitions?: DomainLocalDefinition;
  /**
   *
   * @type {Array<DomainObservation>}
   * @memberof DomainResult
   */
  observations?: Array<DomainObservation>;
  /**
   *
   * @type {Array<DomainProperty>}
   * @memberof DomainResult
   */
  props?: Array<DomainProperty>;
  /**
   *
   * @type {string}
   * @memberof DomainResult
   */
  remarks?: string;
  /**
   *
   * @type {Array<DomainControlsAndObjectives>}
   * @memberof DomainResult
   */
  reviewedControls?: Array<DomainControlsAndObjectives>;
  /**
   *
   * @type {Array<DomainRisk>}
   * @memberof DomainResult
   */
  risks?: Array<DomainRisk>;
  /**
   *
   * @type {string}
   * @memberof DomainResult
   */
  start?: string;
  /**
   *
   * @type {string}
   * @memberof DomainResult
   */
  title?: string;
}